Node Express | D3 | Assignments - Calculator API
Accepts two data in the POST body and based on the url, it returns the results.
Endpoints:
    - /add
    - /multiply
    - /subtract
    - / divide
